Fix sluggish buttons when double-tap is disabled
In order to differentiate between a single tap and a double tap, there
needs to be a delay after the first tap. Otherwise, the action of the
first tap will be executed prematurely.

Unfortunately, that adds a delay to any single tap action. In this case,
the delay was set to 200 milliseconds, which is way more than the
latency of many devices.

The solution is simple – in most cases, people don't even use double
tap, so we can avoid the delay entirely. That's what this commit does.