Treat 0 values as “off”
Zero values meaning “off” is the current behavior of the latest release,
and it is quite useful. For most units, 0 is indeed the absence of
something.

In #1213, it was reported that temperatures are an exception to that.
Indeed! However, the fix made it treat all numeric values as “on”, and
that's not cool!

This commit reverts back the previous behavior, and adds a special case
for temperatures. Now, all units should behave just fine!