Fix `click` propagation
Without this, releasing the mouse click over other elements will trigger
their click events, which can be disastrous. For example, when sliding a
value to 0, it's typical to end up over the navigation drawer, meaning
that the moment the user releases the click, they'll be sent to a
completely different page.

This issue was generally not noticeable on mobile devices, but on desktop
it was quite annoying.