Fix external slider updates
While I'm not following the logic in that code path, the idea seems to
be that it should do an early return if the new value was not processed
yet.

And all that is fine, except that it was using `last_changed` instead of
`last_updated`. For lights, it means that it only reacted to lights
being turned on or off, not when the brightness was changing.

Resolves #1367.